requirements.txt
pathlib
pillow
six
